db: store time difference in int64 to avoid wrap.

+++ db/cache/charts.go
@@ -1254,13 +1254,13 @@ func dailyHashrate(time, chainwork ChartUints) (ChartUints, ChartUints) {
 	rates := make([]uint64, 0, len(time)-1)
 	var dupes int
 	for i, t := range time[1:] {
-		tDiff := t - time[i]
+		tDiff := int64(t - time[i])
 		if tDiff <= 0 {
 			tDiff = aDay
 			dupes += 1
 		}
 		workDiff := chainwork[i+1] - chainwork[i]
-		rates = append(rates, (workDiff)*1e6/tDiff)
+		rates = append(rates, (workDiff)*1e6/uint64(tDiff))
 		times = append(times, t)
 	}
 	if dupes > 0 {
